refactor(embd): move code to shared folder
