+++ embedded/arduino/libs/common/errors.cpp
