+++ embedded/arduino/libs/common/errors.h
