+++ embedded/arduino/libs/io/io_state_machine.h
