+++ embedded/arduino/relay-controller/.gitignore
@@ -1,0 +1,2 @@
+user.mk
+credentials.h
