+++ embedded/arduino/tools/scripts/find_files.py
