+++ embedded/arduino/tools/scripts/remove_files.py
